Adds proper http request to match parse original implementation
